kodecamp-5.0-promotional-task-7/task5-Contact_Manager_API/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
sqlmodel>=0.0.8
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
pydantic>=1.10.2